+++ sjk-core/src/main/java/org/gridkit/jvmtool/SJK.java
@@ -1,5 +1,8 @@
 package org.gridkit.jvmtool;
 
+import java.util.Arrays;
+import java.util.List;
+
 import org.gridkit.jvmtool.cli.CommandLauncher;
 
 public class SJK extends CommandLauncher {
@@ -7,4 +10,13 @@ public class SJK extends CommandLauncher {
     public static void main(String[] args) {
         new SJK().start(args);
     }    
+    
+	@Override
+	protected String[] getModulesUnlockCommand() {
+		return new String[] {"java.base/jdk.internal.perf=ALL-UNNAMED"};
+	}
+
+	protected List<String> getCommandPackages() {
+	    return Arrays.asList("org.gridkit.jvmtool.cmd", "org.gridkit.jvmtool.hflame.cmd");
+	}
 }