+++ sjk-stacktrace/src/main/java/org/gridkit/jvmtool/event/ShieldedEventReader.java
@@ -14,7 +14,7 @@ public class ShieldedEventReader<T extends Event> implements EventReader<T> {
     private static final ErrorHandler THROW_HANDLER = new ErrorHandler() {
 
         @Override
-        public void onException(Exception e) {
+        public boolean onException(Exception e) {
             if (e instanceof RuntimeException) {
                 throw (RuntimeException)e;
             }
@@ -24,16 +24,28 @@ public class ShieldedEventReader<T extends Event> implements EventReader<T> {
         }
     };
 
-    private final EventReader<Event> nested;
+    private static final ErrorHandler SILENT_ERROR_HANDLER = new ErrorHandler() {
+    	
+    	@Override
+    	public boolean onException(Exception e) {
+    		return true; // suppress
+    	}
+    };
+
+    public static <T extends Event> EventReader<T> shield(EventReader<? super T> reader, Class<T> event, boolean suppressErrors) {
+    	return new ShieldedEventReader<T>(reader, event, suppressErrors ? SILENT_ERROR_HANDLER : THROW_HANDLER);
+    }
+    
+    private final EventReader<? extends Event> nested;
     private final Class<T> classFilter;
     private final ErrorHandler errorHandler;
     private T nextEvent;
 
-    public ShieldedEventReader(EventReader<Event> nested, Class<T> classFilter) {
+    public ShieldedEventReader(EventReader<? extends Event> nested, Class<T> classFilter) {
         this(nested, classFilter, THROW_HANDLER);
     }
 
-    public ShieldedEventReader(EventReader<Event> nested, Class<T> classFilter, ErrorHandler errorHandler) {
+    public ShieldedEventReader(EventReader<? extends Event> nested, Class<T> classFilter, ErrorHandler errorHandler) {
         this.nested = nested;
         this.classFilter = classFilter;
         this.errorHandler = errorHandler;
@@ -53,9 +65,12 @@ public class ShieldedEventReader<T extends Event> implements EventReader<T> {
         try {
             while(nested.hasNext()) {
                 Event e = nested.next();
-                if (e instanceof SimpleErrorEvent) {
+                if (e instanceof ErrorEvent) {
                     // forward to error handler
-                    errorHandler.onException(((SimpleErrorEvent) e).exception());
+                    if (errorHandler.onException(((ErrorEvent) e).exception())) {
+                    	// suppress
+                    	continue;
+                    }
                 }
                 if (classFilter.isInstance(e)) {
                     nextEvent = classFilter.cast(e);